Add Basic Login
Project does not have a feature for login, thus allowing
anyone to post on the threads.

Add support for basic login with authentication based on
username. Backend and frontend seem to be communication properly,
with database running fine. Currently working on a database for
posts, and the feature to keep track of authors of posts.